posefusion-client: send as frame with ts & score
The way the bodies are sent is modified such that a message is
sent per frame instead of per body. For example, if three
people are being seen in a frame, an MQTT message containing the
coordinates of the three people is generated and sent all at once.

A timestamp is now also sent in this message corresponding to the
time at which the people MQTT message was generated.

A new selection is made so that only people with a confidence score
above 0.4 is sent. The detected confidence score is also transmitted
for each person detected so that further filtering can be done on the
application side.

Finally, empty frames are sent if no frames were transmitted for over
40 ms. It is used to synchronize the different cameras.